requirements.txt
click
ruamel.yaml
rich 